unix time should be int

+++ sql/ohlc.sql
@@ -37,7 +37,7 @@ ohlc AS (
 
 SELECT
     to_char(period, 'YYYY-MM-DD') || 'T' || to_char(period,'HH24:MI:SSZ') as dt,
-    extract(epoch from period)::numeric as time,
+    extract(epoch from period)::int as time,
     COALESCE(ohlc.open,0)::numeric as open,
     COALESCE(ohlc.high,0)::numeric as high,
     COALESCE(ohlc.low,0)::numeric as low,
